requirements.txt
Brotli==1.0.7
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
Flask==1.1.2
Flask-Assets==2.0
Flask-Compress==1.5.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
pathlib==1.0.1
plotly==4.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
retrying==1.3.3
six==1.14.0
webassets==2.0
Werkzeug==1.0.1